AI_Bot/requirements.txt
Django
gunicorn
python-dotenv
google-genai
# Add any other libraries your code uses